Add files via upload
The dataset given is the details of the care provider, Green Home Help that provide 
temporary staff cover to the local council care workers when they are on leave. The working 
hours covered by a temporary worker is referred to as session. The government Data from the 
individual local councils and Green Home Help are collected of over last 5 years to build a 
data warehouse but cannot be used directly as its quality is poor. The information given 
should be structured in a consistent format and suitable for analysis to build a data 
warehouse.